user listing relationships and finished

+++ evc/app/Http/Controllers/ListingController.php
@@ -43,6 +43,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -53,6 +55,12 @@ class ListingController extends Controller
     } 
 
     public function update(Request $request, Listing $listing) {
+        
+        //ensure current user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -74,7 +82,17 @@ class ListingController extends Controller
 
     //delete a listing
     public function destroy(Listing $listing) {
+        //ensure current user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    //manage listings
+    public function manage() { 
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]); //get current user and gets their listings
     }
 }
